Extract split conversion
