+++ src/main/java/io/split/android/client/storage/splits/SplitEntityConverter.java
@@ -1,0 +1,11 @@
+package io.split.android.client.storage.splits;
+
+import java.util.List;
+
+import io.split.android.client.dtos.Split;
+import io.split.android.client.storage.db.SplitEntity;
+
+public interface SplitEntityConverter {
+
+    List<Split> getFromEntityList(List<SplitEntity> entities);
+}
