+++ src/test/java/io/split/android/client/storage/splits/SqLitePersistentSplitsStorageTest.java
@@ -1,25 +1,16 @@
 package io.split.android.client.storage.splits;
 
-import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyCollection;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
-import io.split.android.client.dtos.Split;
-import io.split.android.client.service.executor.parallel.SplitDeferredTaskItem;
-import io.split.android.client.service.executor.parallel.SplitParallelTaskExecutor;
 import io.split.android.client.storage.db.SplitDao;
 import io.split.android.client.storage.db.SplitEntity;
 import io.split.android.client.storage.db.SplitRoomDatabase;
@@ -29,7 +20,7 @@ public class SqLitePersistentSplitsStorageTest {
     @Mock
     private SplitRoomDatabase mDatabase;
     @Mock
-    private SplitParallelTaskExecutor<List<Split>> mParallelTaskExecutor;
+    private SplitEntityConverter mSplitEntityConverter;
     @Mock
     private SplitDao mSplitDao;
     private SqLitePersistentSplitsStorage mStorage;
@@ -37,71 +28,26 @@ public class SqLitePersistentSplitsStorageTest {
     @Before
     public void setUp() {
         MockitoAnnotations.openMocks(this);
-        when(mParallelTaskExecutor.getAvailableThreads()).thenReturn(2);
-        mStorage = new SqLitePersistentSplitsStorage(mDatabase, mParallelTaskExecutor);
+        mStorage = new SqLitePersistentSplitsStorage(mDatabase, mSplitEntityConverter);
     }
 
     @Test
-    public void getAllUsesParallelExecutorWhenThereAreMoreThan50Splits() {
-        when(mSplitDao.getAll()).thenReturn(getMockEntities(51));
+    public void getAllUsesConverter() {
+        List<SplitEntity> mockEntities = getMockEntities();
+        when(mSplitDao.getAll()).thenReturn(mockEntities);
         when(mDatabase.splitDao()).thenReturn(mSplitDao);
 
         mStorage.getAll();
 
-        verify(mParallelTaskExecutor).execute(anyCollection());
+        verify(mSplitEntityConverter).getFromEntityList(mockEntities);
     }
 
-    @Test
-    public void tasksAreCreatedAccordingToTheAmountOfThreadsAvailable() {
-        ArgumentCaptor<List<SplitDeferredTaskItem<List<Split>>>> argumentCaptor = ArgumentCaptor.forClass(List.class);
-
-        when(mParallelTaskExecutor.getAvailableThreads()).thenReturn(4);
-        List<SplitEntity> mockEntities = getMockEntities(65);
-        when(mSplitDao.getAll()).thenReturn(mockEntities);
-        when(mDatabase.splitDao()).thenReturn(mSplitDao);
-
-        int expectedNumberOfLists = 5;
-
-        mStorage.getAll();
-
-        verify(mParallelTaskExecutor).execute(argumentCaptor.capture());
-        assertEquals(expectedNumberOfLists, argumentCaptor.getValue().size());
-    }
-
-    @Test
-    public void amountOfSplitsEqualsAmountOfEntities() {
-        when(mParallelTaskExecutor.getAvailableThreads()).thenReturn(4);
-        List<SplitEntity> mockEntities = getMockEntities(3);
-        when(mSplitDao.getAll()).thenReturn(mockEntities);
-        when(mDatabase.splitDao()).thenReturn(mSplitDao);
-
-        List<Split> splits = mStorage.getAll();
-
-        assertEquals(3, splits.size());
-    }
-
-    @Test
-    public void amountOfSplitsEqualsAmountOfEntitiesWhenParallel() {
-        when(mParallelTaskExecutor.getAvailableThreads()).thenReturn(2);
-        List<SplitEntity> mockEntities = getMockEntities(3);
-        when(mSplitDao.getAll()).thenReturn(mockEntities);
-        when(mDatabase.splitDao()).thenReturn(mSplitDao);
-        when(mParallelTaskExecutor.execute(any())).thenReturn(
-                Arrays.asList(Collections.singletonList(new Split()),
-                        Collections.singletonList(new Split()),
-                        Collections.singletonList(new Split())));
-
-        List<Split> splits = mStorage.getAll();
-
-        assertEquals(3, splits.size());
-    }
-
-    private List<SplitEntity> getMockEntities(int amount) {
+    private List<SplitEntity> getMockEntities() {
         ArrayList<SplitEntity> entities = new ArrayList<>();
         String jsonTemplate = "{\"name\":\"%s\", \"changeNumber\": %d}";
         long initialChangeNumber = 9999;
 
-        for (int i = 0; i < amount; i++) {
+        for (int i = 0; i < 3; i++) {
             String splitName = "split-" + i;
             SplitEntity entity = new SplitEntity();
             entity.setName(splitName);
